# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Rafiki
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /actividades:
    # our controller name
    x-swagger-router-controller: actividad
    get:
      description: Obtener la lista de actividades
      operationId: getActividades
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResponseActividad"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    post:
      description: Añadir una nueva actividad
      operationId: postActividad
      # movie info to be stored
      parameters:
        - name: data
          description: Atributos de actividad
          in: body
          required: true
          schema:
            $ref: "#/definitions/Actividad"     
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
  /actividades/{id}:
    # our controller name
    x-swagger-router-controller: actividad
    get:
      description: Obtener una actividad
      operationId: getActividad
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResponseActividad"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    put:
      description: Actualizar una actividad
      operationId: updateActividad
      # define the parameters
      parameters:
        - name: id
          description: Id de Actividad
          type: string
          in: path
          required: true
        - name: data
          description: Atributos de actividad
          in: body
          required: true
          schema:
            $ref: "#/definitions/Actividad"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"        
    delete:
      description: Eliminar actividad
      operationId: deleteActividad
      # define the parameters
      parameters:
        - name: id
          description: Id de actividad
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
...
  /asignaturas:
    x-swagger-router-controller: asignatura
    get:
      description: Obtener la lista de asignaturas
      operationId: getAsignaturas
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResponseAsignatura"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    post:
      description: Añadir una nueva asignatura
      operationId: postAsignatura
      # movie info to be stored
      parameters:
        - name: name
          description: Atributos Asignatura
          in: body
          required: true
          schema:
            $ref: "#/definitions/Asignatura"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Asignatura"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
...
  /asignaturas/{id}:
    # our controller name
    x-swagger-router-controller: asignatura
    get:
      description: Obtener un usuario
      operationId: getAsignatura
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResponseAsignatura"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    put:
      description: Actualizar una asignatura
      operationId: updateAsignatura
      # define the parameters
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: data
          description: Propiedades de Asignatura
          in: body
          required: true
          schema:
            $ref: "#/definitions/Asignatura"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Asignatura"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    delete:
      description: Eliminar una Asignatura
      operationId: deleteAsignatura
      # define the parameters
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Errors"
...
  /aprendizajes/{id}:
    # our controller name
    x-swagger-router-controller: objAprendizaje
    get:
      description: Obtener un Objetivo de Aprendizaje
      operationId: getObjAprendizaje
      # define the type of response for Success "200  " and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResponseObjAprendizaje"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    put:
      description: Actualizar un Objetivo de Aprendizaje
      operationId: updateObjAprendizaje
      # define the parameters
      parameters:
        - name: id
          description: ID del Objetivo de Aprendizaje
          type: string
          in: path
          required: true
        - name: data
          description: Propiedades nuevas para el Objetivo de Aprendizaje
          in: body
          required: true
          schema:
            $ref: "#/definitions/ObjAprendizaje"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ObjAprendizaje"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    delete:
      description: un Objetivo de Aprendizaje
      operationId: deleteObjAprendizaje
      # define the parameters
      parameters:
        - name: id
          description: Id del Objetivo de Aprendizaje
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ObjAprendizaje"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
...
definitions:
...
  Asignatura:
    type: object
    properties:
      nombre:
        type: string
        description: Nombre completo de la asignatura
      enfasis:
        type: array
        items: 
          type: string
        description: Descripcion de lo esperado para la asignatura
      horasPedagogicasConJecAnual:
        type: number
        description: Numero de horas anuales con jornada escolar completa
      horasPedagogicasSinJecAnual:
        type: number
        description: Numero de horas anuales sin jornada escolar completa
      horasPedagogicasConJecSemanal:
        type: number
        description: Numero de horas semanales con jornada escolar completa
      horasPedagogicasSinJecSemanal:
        type: number
        description: Numero de horas semanales sin jornada escolar completa
...
  Actividad:
    type: object
    properties:
      profesorAutor:
        type: string
        description: Nombre del autor de la actividad
      anhoAcademico:
        type: number
        description: Año en que se crea la actividad
      semestre:
        type: number
        description: Semestre en que se crea la actividad (1 - 2)
      nivelDificultad:
        type: string
        description: Nivel de dificultad de la actividad
      nivelAprendizaje:
        type: string
        description: Nivel de aprendizaje ('Recordar', 'Comprender', 'Aplicar','Analizar','Evaluar', 'Crear')
      tipoPregunta:
        type: string
        description: Tipo de pregunta ('Enumeracion', 'Conocimiento especifico', 'Ensayo','Verdadero y Falso','Alternativas')
      asignatura:
        type: string
        description: Nombre de la asignatura
      tema:
        type: string
        description: Nombre del tema de la actividad
      subTema:
        type: string
        description: Nombre del subtema de la actividad
      preguntaEnunciado:
        type: string
        description: Enunciado de la pregunta
      preguntaAlternativas:
        type: array
        items: 
          type: string
        description: Alternativas correspondientes al enunciado de la pregunta
      respuesta:
        type: string
        description: Respuesta de referencia 
      respuestaVerdaderoFalso:
        type: boolean
        description: Respuesta de referencia para pregunta de Verdadero y Falso (true - false)
      respuestaAlternativas:
        type: array
        items: 
          type: string
        description: Respuesta(s) de referencia para pregunta de Alternativas
      puntajeTotal:
        type: number
        description: Puntaje total de la pregunta
      retroalimentacion:
        type: array
        items: 
          type: string
        description: Enlaces y/o información complementaria para ser entregada como retroalimentación
...
  Errors:
    type: object
    properties:
      errors:
        type: array
        items:
          type: object
          properties:
            titulo:
              type: string
            detalle:
              type: string
            link:
              type: string
            estado:
              type: string
...
  Response:
    type: object
    properties:
      link:
        type: string
# Objetivos de Aprendizaje
  ObjAprendizaje:
    type: object
    properties:
      numero:
        type: number
        description: numero del objetivo de aprendizaje
      tipo:
        type: string
        description: Tipo de aprendizaje ('Habilidad', 'Conocimiento', 'Actitud')
      descripcion:
        type: string
        description: Informacion respecto a que apunta el objetivo de aprendizaje
  ResponseObjAprendizaje:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items: 
          $ref: "#/definitions/ObjAprendizaje"
      type:
        type: string
  ResponseAsignatura:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items: 
          type: string
          description: Enlaces a las partes que componen el recurso educativo
...
  ResponseActividad:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/Actividad"    
      type:
        type: string     
...